requirements.txt
Django==1.6.2
South==0.8.4
argparse==1.1
dateutils==0.6.6
dj-database-url==0.2.1
dj-static==0.0.5
django-piston==0.2.2.1
gunicorn==18.0
psycopg2==2.5.2
pystache==0.5.3
python-dateutil==2.2
pytz==2013.9
rerun==1.0.19
six==1.5.2
static==1.0.2
wsgiref==0.1.2
